pyproject.toml
[project]
name = "raven"
version = "0.1.0"
description = "Personal LLM training stack with sovereignty over every line"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.0",
    "loguru>=0.7.0",
    "punq>=0.7.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0",
    "tokenizers>=0.20",
    "torch>=2.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "typer>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
